MapSS debugged and works as expected
Former-commit-id: 77565f3f9476c2eb4013adb28977e384941a9b8b

+++ project/Starship/waypoint.cpp
@@ -9,7 +9,7 @@ Vect2d Waypoint::getCentre()
     return this->centre;
 }
 
-double Waypoint::getRadius()
+float Waypoint::getRadius()
 {
     return this->radius;
 }
